+++ src/components/global/header/Docs.mdx
@@ -23,7 +23,10 @@ The primary component is <code>kyn-header-categories</code>. It supports:
 
 ## Properties Reference
 
-<Controls />
+<Controls
+  of={Stories.WithCategorizedNav}
+  exclude={['rootUrl', 'appTitle', 'autoOpenFlyout']}
+/>
 
 ## tabsConfig JSON format
